+++ mulog-core/src/com/brunobonacci/mulog/levels.clj
@@ -1,0 +1,41 @@
+(ns ^{:doc
+      "
+Logging library designed to log data events instead of plain words.
+
+This namespace contains a hierarchy for built-in leveled logging
+as well as a helper to build an event filter.
+"}
+    com.brunobonacci.mulog.levels)
+
+(defonce ^{:doc
+           "Leveled logging hierarchy used by the built-in
+            publishers via the `:level` configuration.
+
+            Users can leverage this hierarchy in custom
+            publishers, or build their own version as they
+            see fit.
+
+            `->filter` is a helper to help implementing custom
+            leveled logging mechanisms.
+            "}
+  default-levels (-> (make-hierarchy)
+                     (derive ::debug    ::verbose)
+                     (derive ::info     ::debug)
+                     (derive ::warning  ::info)
+                     (derive ::error    ::warning)
+                     (derive ::fatal    ::error)))
+
+(defn ->filter
+  "takes a logging level, a `hierarchy` (or `default-levels`)
+  and an event key `k` (or `:mulog/level`) and returns a
+  transformation function which filters events on the given key
+  in the hierarchy"
+  ([level]
+   (or (some-> level (->filter default-levels :mulog/level))
+       identity))
+  ([level hierarchy k]
+   (fn [events]
+     (filter (fn [{e-level k}]
+               (isa? hierarchy e-level level))
+             events))))
+